% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rain_functions.R
\name{get_funnel_area}
\alias{get_funnel_area}
\title{Surface area of the funnel, given its internal diameter}
\usage{
get_funnel_area(funnel_diameter)
}
\arguments{
\item{funnel_diameter}{internal diameter of the funnel (required)}
}
\value{
surface area (numeric). The unit of measure is *funnel_diameter* squared
}
\description{
The larger the funnel diameter, the more rain enters the pluviometer
and the more precise the readings are
}
\details{
It will silently return *NA_real_* for non-numeric and <= 0 funnel diameters
}
\examples{
# Surface area of a 20 cm diameter funnel (314.1593 cm2)
get_funnel_area(20)
}

