% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rain_functions.R
\name{get_pluviometer_factor}
\alias{get_pluviometer_factor}
\title{Get pluviometer factor}
\usage{
get_pluviometer_factor(funnel_area)
}
\arguments{
\item{funnel_area}{The internal surface area of the funnel (in cm2) (required)}
}
\value{
pluviometer factor (numeric). This is a property of the chosen funnel, its units of measure is 1/m2
}
\description{
Pluviometer factor is a property of the funnel you chose and depends on
its surface area. It measure the number of chosen funnels needed to
catch rain in 1 m2 surface area.
}
\examples{
# Get pluviometer factor of a funnel of 314 cm2
get_pluviometer_factor(314)
# To catch rain in 1 m2 one would need 31.84713 funnels of 314 cm2
}

