% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tank_functions.R
\name{get_roof_runoff}
\alias{get_roof_runoff}
\title{Estimate the amount of rain flown into a tank collected on an arbitrary surface}
\usage{
get_roof_runoff(precipitation, precipitation_area, discharge_coef = 0.8,
  first_flush_diverted = 0)
}
\arguments{
\item{precipitation}{the amount of rain fallen in mm (that is l/m2)}

\item{precipitation_area}{the surface area of the roof (in m2)}

\item{discharge_coef}{the discharge coefficient (optional, defaults to 0.8)}

\item{first_flush_diverted}{the amount of water diverted by "first flush devices" (in l, optional, default to 0, i.e. no first flush device installed).}
}
\value{
volume of water runoff (numeric, in l)
}
\description{
Rain tanks typically collect water from house roofs. The amount of roof runoff is very simple:
$(precipitation * precipitation_area * discharge_coef) - first_flush_diverted$
The discharge coefficient determines the proportion of rainfall to the total rainfall, 
that is actually discharged from the roof. An inclined hard roof, with shingles, assume 
a discharge coefficient of 80%, i.e. 0.8, for a green roof assume a a coefficient 
of about 30%.
First flush devices are designed to siphon off the first rainwater that falls on the roof. 
If suc a device is fitted to the rain pipes, the volume of water diverted will have 
to be subtracted from the roof runoff volume.
}
\examples{
# Amount of roof runoff following a 10 mm rain on a 150 m2 roof
get_roof_runoff(10, 150)
# Amount of roof runoff following a 10 mm rain on a 150 m2 green roof,
get_roof_runoff(10, 150, 0.3)
# Amount of roof runoff following a 10 mm rain on a 150 m2 green roof,
# where a first flush device that diverts 50 litres is fitted,
get_roof_runoff(10, 150, 0.3, 50)
}

