% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tank_functions.R
\name{get_roof_runoff_on_same_surface}
\alias{get_roof_runoff_on_same_surface}
\title{Estimate the amount of rain flown into a tank collected for a given surface, discharge coefficient, and first flush volume diverted}
\usage{
get_roof_runoff_on_same_surface(precipitation_area, discharge_coef,
  first_flush_diverted)
}
\arguments{
\item{precipitation_area}{the surface area of the roof}

\item{discharge_coef}{the discharge coefficient}

\item{first_flush_diverted}{the amount of water diverted by "first flush devices" (in l)}
}
\value{
\code{get_roof_runoff} function with arguments \code{precipitation_area} and \code{discharge_coeff} filled
}
\description{
Simple wrapper around \code{get_roof_runoff}
}
\examples{
rainfall_in_tank <- get_roof_runoff_on_same_surface(121, 0.8, 45)
rainfall_in_tank(1)
rainfall_in_tank(100)
}

