% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tank_functions.R
\name{get_tank_charge_on_same_surface}
\alias{get_tank_charge_on_same_surface}
\title{Estimate the amount of rain flown into a tank collected for a given surface and discharge coefficient}
\usage{
get_tank_charge_on_same_surface(precipitation_area, discharge_coef)
}
\arguments{
\item{precipitation_area}{the surface area of the roof}

\item{discharge_coef}{the discharge coefficient determines the proportion of rainfall to the total rainfall, which is actually discharged from the roof (optional, defaults to 0.8, for an inclined hard roof, with shingles)}
}
\value{
\code{get_tank_charge} function with arguments \code{precipitation_area} and \code{discharge_coeff} filled
}
\description{
Simple wrapper around \code{get_tank_charge}
}
\examples{
rainfall_in_tank <- get_tank_charge_on_same_surface(121, 0.8)
rainfall_in_tank(1)
rainfall_in_tank(100)
}

