% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rain_functions.R
\name{get_precipitation_measure}
\alias{get_precipitation_measure}
\title{Convert the amount of water found in the graduated cylinder into l/m2}
\usage{
get_precipitation_measure(water_volume, pluviometer_factor)
}
\arguments{
\item{water_volume}{the volume (in ml) or weight (in g) of the water in the graduated cylinder (required)}

\item{pluviometer_factor}{the factor unique to the funnel chosen (required)}
}
\value{
standard rain measure in l/m2 or mm (numeric)
}
\description{
The amount of water in ml (or g), is converted to l/m2 (or mm) taking into account
the pluviometer factor
}
\examples{
# Found 10 ml of water in a pluviometer whose factor is 31.84713
get_precipitation_measure(10, 31.84713)
}

