% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tank_functions.R
\name{get_tank_charge}
\alias{get_tank_charge}
\title{Estimate the amount of rain flown into a tank collected on an arbitrary surface}
\usage{
get_tank_charge(precipitation, precipitation_area, evaporation_loss = 0)
}
\arguments{
\item{precipitation}{the amount of rain fallen in mm}

\item{precipitation_area}{the surface area of the roof}

\item{evaporation_loss}{mm of rain lost to evaporation on the roof (optional, defaults to 0)}
}
\value{
volume of water collected (numeric, in l)
}
\description{
Rain tanks typically collect water from house roofs.
}
\examples{
# Amount of tank charge following a 10 mm rain on a 150 m2 roof, no evaporation
get_tank_charge(10, 150)
# Amount of tank charge following a 10 mm rain on a 150 m2 roof,
# assuming 2 mm of rain evaporated on hot roof
get_tank_charge(10, 150, 2)
}

